requirements.txt
streamlit>=1.40.0
pandas>=2.2.0
langchain>=0.1.0
langgraph>=0.0.20
langchain-google-genai>=0.0.10  # NEW: Instead of langchain-anthropic
google-generativeai>=0.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.25.0
python-multipart>=0.0.6